requirements.txt
PyYAML==3.11
argparse==1.2.1
requests==2.4.3
wsgiref==0.1.2
